Fix several warnings

+++ libcaf_core/test/simple_timeout.cpp
@@ -104,7 +104,7 @@ CAF_TEST(duration_conversion) {
   duration d1{time_unit::milliseconds, 100};
   std::chrono::milliseconds d2{100};
   duration d3{d2};
-  CAF_CHECK_EQUAL(d1.count, d2.count());
+  CAF_CHECK_EQUAL(d1.count, static_cast<uint64_t>(d2.count()));
   CAF_CHECK_EQUAL(d1, d3);
 }
